added wrapped normal distr, ring.probs
A few new functions. ring.probs, I think, is in line with Sarah's style.
It turns out I used it in the Wolf model too. i.e., calculate the
probabilities of each possible point, multiply them in a few different
dimensions (habitat mediated, correlation mediated) and result is a
probability distribution for turning.