adj is now even faster
For small number of loci <1e4 ... speed ups can be as much as 200x over
original adjacent function in raster package. For 1e6 loci, this
function is only 2.5x faster than adjacent. Note, it now uses either a
matrix approach (small problems) or a data.table approach (big
problems), making the code base a little clunky. It can be cleaned in
the future.