Bug Fix - @$@#$!%
Trying to fix save.R because nothing was saving right. Basically, we
were using the Event queue as a way to determine what to save, plot
etc., but the doEvent method was removing the event from the queue
BEFORE it was being called and completed, so the event queue did not
contain the event that was occurring. Now, the event queue eliminates
the event immediately after it is run, so that the event queue contains
the event being run.