Vignettes - Build! NOTE Path change
The vignettes don't build if they are looking in the system.file()
location of the package. It looks locally where the package is being
built. So, to build the vignettes with R CMD check, the paths have to be
local... this is not totally desirable in all cases, because the user
can not run the code in the Vignettes as a result.