Add clearPlot - a new function
This will clear a plot (surprise) but more importantly, remove all the
hidden metadata associated with that device.  Previously this could only
be done with new=T inside a call to Plot. But, when the call to Plot is
deep within a module, this becomes clunky. clearPlot will just clear it
all.