newModule - rm duplicate "schedule event" suggestions
The suggestion to schedule event occurs twice for each event type...
once inside the doEvent fn and once inside the function called within
the doEvent fn. While this is fine, it was confusing to Yong. It will be
power users who throw scheduleEvent calls all over the place, not
novices.