benchmarking file
Note: Environments have reference semantics. This is relevant because
when we pass a "sim" object that is an environment, any time it gets
changed via, say, sim$landscape <- "hello", we don't have to return the
object. In fact, the functions should use return(invisible()) so that it
is not returned.