new("simEnv") does not work as imagined
because of pass by reference, creating a "new" simEnv does not actually
create a "new" environment... objects tag along between calls. Added a
single line to remove preceding objects stored in a previous simEnv.
Just tested, this only deals with the removal of previous objects... not
with the goal of allowing multiple simEnv objects.