integrate time-awareness into scheduleEvent
This will automatically rescale the scheduleEvent call based on the
timestep of the module.

Plot - bugfix - when only one agent on a map

It created an extent with
a singularity. This now adds padding.

New generic setTimestepUnit

This is set by metadata entry: timestep.
This defines relationships
between the times given in modules. For the
spades call, it can be set
manually to have a different time step for
updates etc.  i.e., for
progress bar, which now shows these units